requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==26.1.0
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colormath==3.0.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
faiss-cpu==1.13.2
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe-silicon==0.9.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.3
open_clip_torch==3.3.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pip-tools==7.5.3
propcache==0.4.1
pyarrow==23.0.1
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pyproject_hooks==1.2.0
protobuf==3.20.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.25
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
